move grouping of wahltagdtos to separate composable
